Changed Structure of Form UI * Changed Department to single select * Changed Role to single select * Show/Hide Fields based on chosen value * Extracted Ticket options as a separate component
